Use generics, instead of big.Float and big.ParseFloat
Bump module name to `v2`, because it has incompatible changes.

Args of `NewThresholds()` must be type of `cmp.Ordered` and it can't be `nil`.

Two new methods added to `Thresholds` type:

```
UseWarning(useMin, useMax bool)
UseCritical(useMin, useMax bool)
```

`useMin` defines is `WarningMin` or `CriticalMin` defined or not. `useMax` does
the same for `WarningMax` or `CriticalMax`. So instead of calling
`NewThresholds()` with `nil`s, now we should use `UseWarning()` and/or
`UseCritical()`.